--- client/src/action.js
@@ -1,6 +1,0 @@
-export const auth = (userInfo) =>{
-    return{
-        type: 'AUTH',
-        payload: userInfo,
-    };
-};
