--- client/src/jobsList.js
@@ -1,4 +1,0 @@
-const jobs=[
-    {'id':112, 'description': "kayak", 'orgin': "Seattle, Wa", 'Destination':"Denver, Co"},
-    {},
-]
